backend/python/requirements.txt
Django==4.0.0
django-cors-headers==3.13.0
django-extensions==3.1.5
djangorestframework==3.13.1
psycopg2==2.9.3
PyJWT==2.4.0
six==1.16.0
